Fix precision error by using long double

+++ Mathematics/Inversion Probability.cpp
@@ -5,7 +5,7 @@ typedef long long ll;
 const int maxN = 101;
 
 int N, r[maxN];
-double ans;
+long double ans;
 
 int f(int x){
     return x * (x-1) / 2;
@@ -17,8 +17,8 @@ int main(){
         scanf("%d", &r[i]);
         for(int j = 0; j < i; j++){
             int tot = (r[j] <= r[i]) ? f(r[j]) : (f(r[i]) + (r[j]-r[i]) * r[i]);
-            ans += (double) tot / (r[i] * r[j]);
+            ans += (long double) tot / (r[i] * r[j]);
         }
     }
-    printf("%.6f\n", ans);
+    printf("%.6Lf\n", ans);
 }
